Update README.md
🚀 Project Milestone: Static Resume Builder 🎉
Just completed and deployed my first milestone project—a Static Resume Builder! Built using HTML, CSS, and TypeScript, this tool provides an easy and sleek way to generate professional resumes. Proudly published on GitHub and live on Vercel for seamless access. It's been a great hands-on experience with front-end development, and I'm excited for what's next!